Backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
motor>=2.5.0
pymongo>=3.12.0
bcrypt>=3.2.0
python-jose>=3.3.0
passlib>=1.7.4
pandas>=1.3.0
sendgrid>=6.8.0
python-multipart>=0.0.5
email-validator>=1.1.3
pydantic[email]>=1.8.2
python-jose[cryptography]>=3.3.0
bson>=0.5.10